refactor(handler): extract DoReadOne from ReadOneWeb
